+++ CHANGELOG.md
@@ -25,6 +25,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Added
 
 - Added Skill and Attribute names to dice
+- Added `SwadeEntityTweaks` class to game object as `game.swade.SwadeEntityTweaks`.
 
 ### Changed